Added the forest tileset, and made a few small bug fixes.

+++ Delve/ChangeMapViewController.m
@@ -130,7 +130,7 @@
 -(void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
 	//set the info text
-	[self setInfoText:indexPath.row];
+	[self setInfoText:(int)indexPath.row];
 }
 
 @end
